// src/lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  TECHNICIAN: 'TECHNICIAN',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PoleStatus = {
  ACTIVE: 'ACTIVE',
  FAULTY: 'FAULTY',
  UNDER_MAINTENANCE: 'UNDER_MAINTENANCE',
  DECOMMISSIONED: 'DECOMMISSIONED'
} as const

export type PoleStatus = (typeof PoleStatus)[keyof typeof PoleStatus]


export const FaultType = {
  NO_POWER: 'NO_POWER',
  FLICKERING: 'FLICKERING',
  DAMAGED_FIXTURE: 'DAMAGED_FIXTURE',
  VANDALISM: 'VANDALISM',
  OTHER: 'OTHER'
} as const

export type FaultType = (typeof FaultType)[keyof typeof FaultType]


export const ReportStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const WorkOrderStatus = {
  PENDING: 'PENDING',
  ASSIGNED: 'ASSIGNED',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  CANCELLED: 'CANCELLED'
} as const

export type WorkOrderStatus = (typeof WorkOrderStatus)[keyof typeof WorkOrderStatus]